+++ requirements.txt
@@ -1,49 +1,2 @@
-Babel==1.3
-Mako==1.0.1
-MarkupSafe==0.23
-PasteDeploy==1.5.2
-SPARQLWrapper==1.6.4
-SQLAlchemy==1.0.7
-WebHelpers2==2.0
-WebOb==1.4.1
-WebTest==2.0.18
-alembic==0.7.6
-argparse==1.2.1
-beautifulsoup4==4.3.2
-clld==1.1.0
-clldmpg==1.0.0
-colander==1.0
-ecdsa==0.13
-future==0.14.3
-html5lib==0.999999
-iso8601==0.1.10
-isodate==0.5.1
-nameparser==0.3.4
-paginate==0.4.2
-path.py==7.3
-pathlib==1.0.1
-purl==1.0.3
-pyparsing==2.0.3
-pyramid==1.6a2
-pyramid-exclog==0.7
-pyramid-mako==1.0.2
-pyramid-tm==0.11
-python-dateutil==2.4.2
-pytz==2015.2
-rdflib==4.2.0
-repoze.lru==0.6
-requests==2.6.2
-six==1.9.0
-transaction==1.4.3
-translationstring==1.3
-venusian==1.0
-virtualenv==13.1.0
-waitress==0.8.9
-webassets==0.10.1
-wsgiref==0.1.2
-xlrd==0.9.3
-xlwt-future==0.8.0
-yuicompressor==2.4.8
-zope.deprecation==4.1.2
-zope.interface==4.1.2
-zope.sqlalchemy==0.7.6
+# development environment: install in development mode
+-e .[dev,test]
