+++ setup.cfg
@@ -6,6 +6,12 @@ addopts =
     --cov=sails
     --cov-report term-missing
 
+[coverage:run]
+source = sails
+
+[coverage:report]
+show_missing = true
+
 [compile_catalog]
 directory = sails/locale
 domain = clld
